backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-backend"
version = "0.1.0"
description = "Backend for the KBF paper authoring system"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "psycopg[binary]>=3.2.0",
  "python-docx>=1.1.2",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.20",
  "redis>=5.2.0",
  "requests>=2.32.0",
  "rq>=1.16.2",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]
